#23 refactor: 로직 컴포넌트, 뷰 컴포넌트 분리 - 별도 뷰 컴포넌트 생성, 로직 컴포넌트는 하나의 뷰 컴포넌트만 반환하도록 수정
